gradle config refactoring

+++ clap-server/clap-server-service/src/main/java/com/noveogroup/clap/service/file/FileServiceImpl.java
@@ -66,9 +66,11 @@ public class FileServiceImpl implements FileService {
 
     @Override
     public boolean removeFile(final String path) {
-        File file = new File(path);
-        if (file.exists()) {
-            return file.delete();
+        if(path != null){
+            final File file = new File(path);
+            if (file.exists()) {
+                return file.delete();
+            }
         }
         return false;
     }
